refactor: move the engine into screen/ and its output vocabulary into render/
The last of the moves, and the one that breaks the module's only directory-level
cycle -- by placement rather than by an API change.

The cycle was never in the headers. RenderBuffer.hpp names Terminal zero times,
and StatusLineBuilder.hpp only forward-declares it. It came entirely from filing
the BUILDERS together with the DATA they produce: RenderBufferBuilder holds a
Terminal const* and calls fifteen of its members, plus seven of Screen's --
at(), multicellBlockAt(), cursor() and the four cell accessors. StatusLineBuilder
is the same shape, its VTSerializer reading fourteen Terminal members across some
three hundred lines.

So the split is builders from vocabulary, not rendering from engine:

  render/   RenderBuffer -- what the rasterizer consumes, and nothing else.
            Four files in vtrasterizer include it; it is a published boundary,
            which is what earns a directory its place at one unit.
  screen/   Screen, Terminal, and everything that walks them: Settings, Cursor,
            Viewport, RenderBufferBuilder, StatusLineBuilder.

  grep -rn '#include <vtbackend/screen/' src/vtbackend/render/

is empty, and the graph is now acyclic: core/ -> grid/ -> vt/ -> render/ -> screen/.

The alternative was to hand the builders a frame-context struct so they could stay
in render/. It was drafted and rejected: severing the edge needs all twenty-two
members abstracted, which puts nine indirect calls on the per-cell path -- and
that path runs under _stateMutex, which Terminal.hpp itself documents as heavily
contended and which the parser thread holds for the full duration of an output
burst. Trading VT-thread latency for a directory boundary is the wrong way round.
Placement costs nothing and says the same thing.

Cursor and Viewport come here rather than to core/ and grid/ respectively.
Cursor includes Charset and is the DECSC/DECRC saved record, named only by Screen
and Terminal. Viewport includes Screen and is constructed from a Terminal&; it has
no Grid include at all, being a view onto the terminal rather than onto the grid.

What is left at the top of src/vtbackend/ is Logging.hpp, the two tools, and the
test scaffolding that the next commit moves.
