+++ src/contour/session/FontControl.cpp
@@ -5,7 +5,7 @@
 #include <contour/session/Logging.hpp>
 #include <contour/session/TerminalSession.hpp>
 
-#include <vtbackend/Terminal.hpp>
+#include <vtbackend/screen/Terminal.hpp>
 
 #include <vtrasterizer/Renderer.hpp>
