+++ src/contour/session/SessionFactory.hpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 #pragma once
 
-#include <vtbackend/Terminal.hpp>
 #include <vtbackend/core/Primitives.hpp>
+#include <vtbackend/screen/Terminal.hpp>
 
 #include <vtpty/Process.hpp>
 #include <vtpty/Pty.hpp>
