+++ src/vtbackend/graphics/GoodImageProtocol_test.cpp
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <vtbackend/MockTerm.hpp>
-#include <vtbackend/Screen.hpp>
 #include <vtbackend/TestHelpers.hpp>
 #include <vtbackend/core/Image.hpp>
 #include <vtbackend/graphics/MessageParser.hpp>
+#include <vtbackend/screen/Screen.hpp>
 
 #include <crispy/Base64.hpp>
