+++ src/vtbackend/input/vi/Selector_test.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <vtbackend/MockTerm.hpp>
-#include <vtbackend/Screen.hpp>
 #include <vtbackend/input/vi/Selector.hpp>
+#include <vtbackend/screen/Screen.hpp>
 
 #include <catch2/catch_test_macros.hpp>
