+++ src/vtbackend/screen/StatusLineBuilder_test.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/StatusLineBuilder.hpp>
+#include <vtbackend/screen/StatusLineBuilder.hpp>
 
 #include <catch2/catch_test_macros.hpp>
