+++ src/vtbackend/shell/Folding_test.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <vtbackend/MockTerm.hpp>
-#include <vtbackend/StatusLineBuilder.hpp>
-#include <vtbackend/Terminal.hpp>
+#include <vtbackend/screen/StatusLineBuilder.hpp>
+#include <vtbackend/screen/Terminal.hpp>
 #include <vtbackend/shell/Folding.hpp>
 
 #include <crispy/Utils.hpp>
